requirements.txt
Django==3.1.3
psycopg2==2.8.6
django-allauth
django-crispy-forms
django-invitations
django-import-export
wordcloud
coverage
pandas